% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R, R/teatimer-package.r
\docType{package}
\name{teatimer}
\alias{teatimer}
\alias{teatimer}
\alias{teatimer-package}
\title{teatimer}
\usage{
teatimer(time = 1, minutes = TRUE, sound = 0)
}
\arguments{
\item{time}{A positive number, the amount of time to wait}

\item{minutes}{If TRUE, time is in minutes. If FALSE, time is in seconds.}

\item{sound}{character string or number specifying what sound to be played by either specifying
one of the built in sounds, specifying the path to a wav file or specifying an url. The default is 0.
See the documentation for \code{\link[beepr:beep]{beep}}.}
}
\description{
teatimer is a timer that, by default, runs for however many minutes specified and then
plays an sound when the time is up. It can be set to run in seconds rather than minutes
and the sound can also be specified.

teatimer.
}
\examples{
\dontrun{
teatimer(time = 0, minutes = TRUE, sound = 1)
}

}
